Harden AES decryption; cap Reference number and generation.
- crypto: constant-time PKCS#7 unpad in DecryptBytes and decryptReader,
  closing a padding oracle that previously only inspected the trailing
  length byte.  Add a regression test that flips a CBC byte to corrupt
  the pad region while leaving the length byte intact.
- crypto: constant-time compares for U/O password validation (R 2-6).
- crypto: /Perms (Algorithm 10) now fills bytes 12-15 with random data
  as required by §7.6.4.4.9 step (e), and Algorithm 13 validation folds
  every field check into a single accumulator.  This adds a previously
  missing check that bytes 4-7 of the decrypted Perms equal 0xFF, per
  §7.6.4.4.5 step (a) — a behavior tightening: producers that left those
  bytes unset will now fail authentication (Acrobat and this library
  always set them, so impact should be ~zero in practice).
- Cap object numbers at 2^24-1 (PDF 32000-2 §7.6.3.2 step (b) only mixes
  the low 3 bytes into per-object encryption keys, so higher values
  would silently collide).  NewReference panics on out-of-range input
  and Writer.Alloc panics if the ceiling is reached.
- Narrow Reference.Generation to uint16 — §7.5.4 caps generations at
  65535, and §7.6.3.2 only mixes 2 bytes into key derivation.  Readers
  silently drop xref-stream entries whose generation exceeds the cap.
- New constants maxXRefSize (1<<24) and maxGeneration (65535) replace
  scattered math.MaxUint32 checks across scanner, sequential, and xref.

+++ sequential_test.go
@@ -76,7 +76,7 @@ trailer
 
 	type objInfo struct {
 		pos int64
-		gen uint32
+		gen uint16
 	}
 	objs := make(map[uint32]objInfo)
 	for _, sect := range info.Sections {
